Reorganizes folders out of app to make them more global
