+++ src/app/api/generate/client.ts
@@ -1,6 +1,6 @@
 // src/app/api/generate/client.ts
 
-import { GenerateResponse } from '@/app/types/api'
+import { GenerateResponse } from '@/types/api'
 
 export async function generateSentence(sentence: string): Promise<GenerateResponse> {
   const res = await fetch('/api/generate', {
