+++ src/app/components/ResponseDisplay.test.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 import { render, screen } from '@testing-library/react'
 import { describe, expect, test } from 'vitest'
 import '@testing-library/jest-dom'
-import { useSentenceStore } from '@/app/stores/useSentenceStore'
+import { useSentenceStore } from '@/stores/useSentenceStore'
 import ResponseDisplay from './ResponseDisplay'
 
 describe('ResponseDisplay', () => {
