+++ src/app/components/SentenceInput.test.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 import { render, screen, fireEvent } from '@testing-library/react'
 import { beforeEach, describe, expect, test, vi } from 'vitest'
 import '@testing-library/jest-dom'
-import { useSentenceStore } from '@/app/stores/useSentenceStore'
+import { useSentenceStore } from '@/stores/useSentenceStore'
 import SentenceInput from './SentenceInput'
 
 // Reset store before each test
